% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clus_explorer_utils_reports.R
\name{chart_heatmap}
\alias{chart_heatmap}
\title{Function that returns ggplot heatmap chart}
\usage{
chart_heatmap(
  data,
  x,
  y,
  z,
  legend_position = "right",
  title = "",
  subtitle = "",
  caption = "",
  n.dodge = NULL,
  check.overlap = FALSE
)
}
\arguments{
\item{data}{Data to be plotted}

\item{x}{X coordinate}

\item{y}{Y coordinate}

\item{z}{Z coordinate}

\item{legend_position}{Legend position ('top', 'right', 'bottom', 'left')}

\item{title}{Chart title}

\item{subtitle}{Chart subtitle}

\item{caption}{Chart caption}

\item{n.dodge}{Dodge parameter}

\item{check.overlap}{Whether to check for overlapping labels}
}
\value{

}
\description{
Function that returns ggplot heatmap chart
}
