% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clus_explorer_utils_reports.R
\name{chart_line_faceted}
\alias{chart_line_faceted}
\title{Reusable function to create a line chart}
\usage{
chart_line_faceted(
  data,
  x_var,
  y_var,
  color_var,
  facet_chart = FALSE,
  facet_vars = NULL,
  xlab = "",
  ylab = "",
  is_plotly = FALSE,
  legend_position = "bottom",
  add_x_intercept = FALSE,
  x_intercept = 0,
  add_y_intercept = FALSE,
  y_intercept = 0
)
}
\arguments{
\item{data}{Data to be plotted}

\item{x_var}{X-axis variable}

\item{y_var}{Y-axis variable}

\item{color_var}{Colour variable}

\item{facet_chart}{Should the chart be faceted}

\item{facet_vars}{Facet variables}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{is_plotly}{Should the chart object be converted to plotly object}
}
\value{

}
\description{
Reusable function to create a line chart
}
