% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clus_explorer_utils_server.R
\name{getAvailableStudyAreas}
\alias{getAvailableStudyAreas}
\title{Get available study areas to populate the "Area of interest" dropdown}
\usage{
getAvailableStudyAreas()
}
\value{

}
\description{
Get available study areas to populate the "Area of interest" dropdown
}
