% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clus_explorer_utils_server.R
\name{getRasterQuery}
\alias{getRasterQuery}
\title{Get the rasters from a PostGIS table}
\usage{
getRasterQuery(srcRaster)
}
\arguments{
\item{srcRaster}{A character string specifying a PostgreSQL schema and table/view name holding the geometry (e.g., name = c("schema","table"))}
}
\value{

}
\description{
Get the rasters from a PostGIS table
}
