% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clus_explorer_utils_server.R
\name{getSpatialQuery}
\alias{getSpatialQuery}
\title{Get the results of a spatial query from the PostgreSQL database}
\usage{
getSpatialQuery(sql)
}
\arguments{
\item{sql}{The SQL query}
}
\value{

}
\description{
Get the results of a spatial query from the PostgreSQL database
}
