% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mod_chart_area.R
\name{mod_chart_area_ui}
\alias{mod_chart_area_ui}
\alias{mod_chart_area_server}
\title{mod_chart_area_ui and mod_chart_area_server}
\usage{
mod_chart_area_ui(id, chart_height = "450px")

mod_chart_area_server(
  id,
  dataset,
  x,
  y,
  trace_name = "",
  chart_type = "scatter",
  stackgroup = "one",
  chart_mode = "markers+lines",
  traces = list(),
  marker_color = "#F29B9B",
  line_shape = "linear",
  line_fill = "none",
  line_fillcolor = "rgba(160, 129, 217, 0.25)",
  x_axis_tick_format = "",
  x_axis_nticks = 0,
  legend_y = -0.1,
  x_axis_tick_prefix = "",
  y_axis_tick_format = "",
  y_axis_tick_prefix = ""
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module for plotly stacked area chart.
}
\keyword{internal}
