% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mod_chart_line.R
\name{mod_mod_chart_line_server}
\alias{mod_mod_chart_line_server}
\title{mod_chart_line Server Function for plotly line chart}
\usage{
mod_mod_chart_line_server(
  id,
  dataset,
  x,
  y,
  trace_name = "",
  chart_type = "scatter",
  chart_mode = "marker+lines",
  traces = list(),
  marker_color = "#F29B9B",
  line_shape = "linear",
  line_fill = "none",
  line_fillcolor = "rgba(160, 129, 217, 0.25)",
  x_axis_tick_format = "string",
  x_axis_nticks = 0,
  legend_y = -0.15,
  x_axis_tick_prefix = "",
  y_axis_tick_format = "string",
  y_axis_tick_prefix = "",
  shapes = list(),
  add_annotations = FALSE,
  annotation_text = c(),
  annotation_x = c(),
  annotation_y = c(),
  annotation_legend = FALSE,
  add_loess = FALSE,
  loess_frequency = 1,
  loess_mean_colour = "#c9c9c9",
  loess_ribbon_colour = "rgba(96, 96, 96, 0.15)"
)
}
\arguments{
\item{dataset}{dataset to be plotted}

\item{x}{X coordinate variable}

\item{y}{Y coordinate variable}

\item{trace_name}{Trace name}

\item{chart_type}{Chart type}

\item{chart_mode}{Chart mode}

\item{traces}{Traces}

\item{marker_color}{Marker colour}

\item{line_shape}{Line shape}

\item{line_fill}{Line fill}

\item{line_fillcolor}{Line fill colour}

\item{x_axis_tick_format}{X axis tick format}

\item{x_axis_nticks}{X axis nticks}

\item{legend_y}{Legend Y}

\item{x_axis_tick_prefix}{X axis tick prefix}

\item{y_axis_tick_format}{Y axis tick format}

\item{y_axis_tick_prefix}{Y axis tick prefix}

\item{shapes}{Shapes to add to the plot}

\item{add_annotations}{Add annotations}

\item{annotation_text}{Annotation text}

\item{annotation_x}{Annotation text X coordinate}

\item{annotation_y}{Annotation text Y coordinate}

\item{annotation_legend}{Show annotation legend or not}

\item{add_loess}{Add LOESS smoothing or not}

\item{loess_frequency}{LOESS smoothing frequency}

\item{loess_mean_colour}{LOESS mean colour}

\item{loess_ribbon_colour}{LOESS ribbon colour}

\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}
}
\description{
mod_chart_line Server Function for plotly line chart
}
