% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clus_explorer_utils_ui.R
\name{valueBoxSpark}
\alias{valueBoxSpark}
\title{Value box}
\usage{
valueBoxSpark(
  value = NULL,
  title = NULL,
  sparkobj = NULL,
  subtitle = NULL,
  info = NULL,
  icon = NULL,
  color = "aqua",
  width = 4,
  href = NULL
)
}
\arguments{
\item{value}{Value}

\item{title}{Title}

\item{sparkobj}{Object}

\item{subtitle}{Subtitle}

\item{info}{Info}

\item{icon}{Icon}

\item{color}{Box colour}

\item{width}{Box width}

\item{href}{Box href}
}
\value{

}
\description{
Value box
}
