% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clus_explorer_utils_reports.R
\name{chart_line_faceted}
\alias{chart_line_faceted}
\title{Reusable function to create a line chart}
\usage{
chart_line_faceted(
  data,
  x_var,
  y_var,
  color_var,
  facet_chart = FALSE,
  facet_vars = NULL,
  facet_scales = "free",
  facet_nrow = 3,
  facet_ncol = 3,
  xlab = "",
  ylab = "",
  is_plotly = FALSE,
  add_x_intercept = FALSE,
  x_intercept = 0,
  add_y_intercept = FALSE,
  y_intercept = 0,
  legend_position = "bottom",
  strip.position = "top",
  labeller_data = c(),
  height = 600
)
}
\arguments{
\item{data}{Data to be plotted}

\item{x_var}{X-axis variable}

\item{y_var}{Y-axis variable}

\item{color_var}{Colour variable}

\item{facet_chart}{Should the chart be faceted}

\item{facet_vars}{Facet variables}

\item{facet_scales}{Should scales be fixed ("fixed", the default), free ("free"), or free in one dimension ("free_x", "free_y")?}

\item{facet_nrow}{Number of facet rows}

\item{facet_ncol}{Number columns.}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{is_plotly}{Should the chart object be converted to plotly object}

\item{add_x_intercept}{Whether to add vertical intercept line}

\item{x_intercept}{Position of vertical intercept line}

\item{add_y_intercept}{Whether to add horizontal intercept line}

\item{y_intercept}{Position of vertical horizontal line}

\item{legend_position}{Legend position}

\item{labeller_data}{Facet labeller data}

\item{height}{Chart height}

\item{strip_position}{Facet strip position}
}
\value{

}
\description{
Reusable function to create a line chart
}
