% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clus_explorer_utils_server.R
\name{getTableQuery}
\alias{getTableQuery}
\title{Get the results of an SQL query from the database}
\usage{
getTableQuery(sql, params = list())
}
\arguments{
\item{sql}{The SQL query}

\item{params}{The SQL query parameters}
}
\value{

}
\description{
Get the results of an SQL query from the database
}
